Move data provider to data package
